Securisation de project & schoolYear

+++ src/Controller/ProjectController.php
@@ -4,11 +4,13 @@ namespace App\Controller;
 
 use App\Entity\Project;
 use App\Form\ProjectType;
+use App\Form\ProjectForStudentType;
 use App\Repository\ProjectRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 
 /**
@@ -21,14 +23,26 @@ class ProjectController extends AbstractController
      */
     public function index(ProjectRepository $projectRepository): Response
     {
-        // $this->denyAccessUnlessGranted([
-        //     'ROLE_ADMIN',
-        //     'ROLE_TEACHER',
-        // ]);
+        if (!$this->isGranted('ROLE_ADMIN') //filtre des users
+            && !$this->isGranted('ROLE_TEACHER')
+            && !$this->isGranted('ROLE_STUDENT')
+            && !$this->isGranted('ROLE_CLIENT')
+        ) {
+            throw new AccessDeniedException();
+        }
 
+        if ($this->isGranted('ROLE_STUDENT')
+            || $this->isGranted('ROLE_CLIENT')
+        ) {
+            // les projets de l'utilisateur
+            $projects = $this->getUser()->getProjects();
+        } else {
+            // tout les projets
+            $projects = $projectRepository->findAll();
+        }
 
         return $this->render('project/index.html.twig', [
-            'projects' => $projectRepository->findAll(),
+            'projects' => $projects,
         ]);
     }
 
@@ -37,6 +51,12 @@ class ProjectController extends AbstractController
      */
     public function new(Request $request): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN') //filtre des users
+        && !$this->isGranted('ROLE_TEACHER')
+        ) {
+            throw new AccessDeniedException();
+        }
+
         $project = new Project();
         $form = $this->createForm(ProjectType::class, $project);
         $form->handleRequest($request);
@@ -60,6 +80,24 @@ class ProjectController extends AbstractController
      */
     public function show(Project $project): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')
+        && !$this->isGranted('ROLE_TEACHER')
+        && !$this->isGranted('ROLE_STUDENT')
+        && !$this->isGranted('ROLE_CLIENT')
+        ) {
+            throw new AccessDeniedException();
+        }
+
+        if($this->isGranted('ROLE_STUDENT')
+            || $this->isGranted('ROLE_CLIENT')
+        ) {
+            $projects = $this->getUser()->getProjects(); // on recupere tt les projects d'utilisateur
+
+            if (!$projects->contains($project)) {   // si le project demandé ne fait pas partie des project de l'utilisateur
+                throw new AccessDeniedException();  // acces denied
+            }
+        }
+
         return $this->render('project/show.html.twig', [
             'project' => $project,
         ]);
@@ -70,7 +108,66 @@ class ProjectController extends AbstractController
      */
     public function edit(Request $request, Project $project): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')
+        && !$this->isGranted('ROLE_TEACHER')
+        ) {
+            throw new AccessDeniedException();
+        }
+
+        if($this->isGranted('ROLE_STUDENT')) {
+            $projects = $this->getUser()->getProjects(); // on recupere tt les projects d'utilisateur
+
+            if (!$projects->contains($project)) {   // si le project demandé ne fait pas partie des project de l'utilisateur
+                throw new AccessDeniedException();  // acces denied
+            }
+        }
+
         $form = $this->createForm(ProjectType::class, $project);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $this->getDoctrine()->getManager()->flush();
+
+            return $this->redirectToRoute('project_index');
+        }
+
+        return $this->render('project/edit.html.twig', [
+            'project' => $project,
+            'form' => $form->createView(),
+        ]);
+    }
+
+    /**
+     * @Route("/{id}/edit-for-student", name="project_edit_for_student", methods={"GET","POST"})
+     */
+    public function editForStudent(Request $request, Project $project): Response
+    {
+        // une négation de « ou »
+        // if (!(
+        //         $this->isGranted('ROLE_ADMIN')
+        //         || $this->isGranted('ROLE_TEACHER')
+        //         || $this->isGranted('ROLE_STUDENT')
+        // )){
+        //     throw new AccessDeniedException();
+        // }
+
+        // des négations de « et »
+        if (!$this->isGranted('ROLE_ADMIN')
+            && !$this->isGranted('ROLE_TEACHER')
+            && !$this->isGranted('ROLE_STUDENT')
+        ) {
+            throw new AccessDeniedException();
+        }
+
+        if ($this->isGranted('ROLE_STUDENT')) {
+            $projects = $this->getUser()->getProjects();
+
+            if (!$projects->contains($project)) {
+                throw new AccessDeniedException();
+            }
+        }
+
+        $form = $this->createForm(ProjectForStudentType::class, $project);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -90,6 +187,12 @@ class ProjectController extends AbstractController
      */
     public function delete(Request $request, Project $project): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')
+        && !$this->isGranted('ROLE_TEACHER')
+        ) {
+            throw new AccessDeniedException();
+        }
+
         if ($this->isCsrfTokenValid('delete'.$project->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($project);
